+++ tests/FSharp.Data.GraphQL.Benchmarks/Program.fs
@@ -4,14 +4,14 @@ open BenchmarkDotNet.Running
 open FSharp.Data.GraphQL.AsyncValBenchmark
 open FSharp.Data.GraphQL.ParsingBenchmark
 open FSharp.Data.GraphQL.ExecutionBenchmark
-open FSharp.Data.GraphQL.MiddlewaresBenchmark
+open FSharp.Data.GraphQL.MiddlewareBenchmark
 
 let defaultSwitch () = 
     BenchmarkSwitcher [| 
         typeof<AsyncValBenchmark>
         typeof<SimpleExecutionBenchmark>
         typeof<ParsingBenchmark>
-        typeof<SimpleExecutionWithMiddlewaresBenchmark>
+        typeof<SimpleExecutionWithMiddlewareBenchmark>
     |]
 
 [<EntryPoint>]